requirements.txt
beautifulsoup4==4.6.0
lxml==4.1.0
PyYAML==3.12
